Allow overriding base order for react-style styles
This feature intended for increasing specificity of styles added using
react-style, for example because react-style is combined with an
external style sheet that needs to be overridden.

Another way of setting base specificity would be adding a corresponding
argument to the ReactStyle call.  However this would mean that when
specifying multiple styles in an array, you can't rely on later styles
overriding earlier ones anymore.  Therefore this is implemented as a
per-element property.